+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 - chore - configure renovate
 - chore - remove dependabot config (duplicates renovate)
 - chore - remove safety (duplicates sonarqube)
+- change - swap pebble for mpire, kill procs on run timeout
 
 ## 0.197.0
 - Fix - before all hooks are now executed